pyproject.toml
[project]
name = "equity-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.1",
    "fpdf>=1.7.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "litellm[proxy]>=1.82.4",
    "mcp[cli]>=1.26.0",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "streamlit>=1.55.0",
    "textblob>=0.19.0",
    "yfinance>=1.2.0",
]
